// src/client/UI/game/_types/IKeyLayout.ts
/** A keyboard layout */
export type IKeyLayout = IKey[][];

/** A single key on the keyboard */
export type IKey =
    | "a"
    | "b"
    | "c"
    | "d"
    | "e"
    | "f"
    | "g"
    | "h"
    | "i"
    | "j"
    | "k"
    | "l"
    | "m"
    | "n"
    | "o"
    | "p"
    | "q"
    | "r"
    | "s"
    | "t"
    | "u"
    | "v"
    | "w"
    | "x"
    | "y"
    | "z"
    | "enter"
    | "backspace";